src/golang.org/x/net/html/testdata/webkit/tests22.dat

#data
<q><b><big><em><strong><div>X</q>
#errors
#document
| <html>
|   <head>
|   <body>
|     <q>
|       <b>
|         <big>
|           <em>
|             <strong>
|     <big>
|       <em>
|         <strong>
|           <div>
|             <q>
|               "X"

#data
<q><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8>A</q>
#errors
#document
| <html>
|   <head>
|   <body>
|     <q>
|       <b>
|     <b>
|       <div>
|         id="1"
|         <q>
|         <div>
|           id="2"
|           <q>
|           <div>
|             id="3"
|             <q>
|             <div>
|               id="4"
|               <q>
|               <div>
|                 id="5"
|                 <q>
|                 <div>
|                   id="6"
|                   <q>
|                   <div>
|                     id="7"
|                     <q>
|                     <div>
|                       id="8"
|                       <q>
|                         "A"

#data
<q><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8><div id=9>A</q>
#errors
#document
| <html>
|   <head>
|   <body>
|     <q>
|       <b>
|     <b>
|       <div>
|         id="1"
|         <q>
|         <div>
|           id="2"
|           <q>
|           <div>
|             id="3"
|             <q>
|             <div>
|               id="4"
|               <q>
|               <div>
|                 id="5"
|                 <q>
|                 <div>
|                   id="6"
|                   <q>
|                   <div>
|                     id="7"
|                     <q>
|                     <div>
|                       id="8"
|                       <q>
|                         <div>
|                           id="9"
|                           "A"

#data
<q><b><div id=1><div id=2><div id=3><div id=4><div id=5><div id=6><div id=7><div id=8><div id=9><div id=10>A</q>
#errors
#document
| <html>
|   <head>
|   <body>
|     <q>
|       <b>
|     <b>
|       <div>
|         id="1"
|         <q>
|         <div>
|           id="2"
|           <q>
|           <div>
|             id="3"
|             <q>
|             <div>
|               id="4"
|               <q>
|               <div>
|                 id="5"
|                 <q>
|                 <div>
|                   id="6"
|                   <q>
|                   <div>
|                     id="7"
|                     <q>
|                     <div>
|                       id="8"
|                       <q>
|                         <div>
|                           id="9"
|                           <div>
|                             id="10"
|                             "A"

#data
<cite><b><cite><i><cite><i><cite><i><div>X</b>TEST
#errors
Line: 1 Col: 6 Unexpected start tag (cite). Expected DOCTYPE.
Line: 1 Col: 46 End tag (b) violates step 1, paragraph 3 of the adoption agency algorithm.
Line: 1 Col: 50 Expected closing tag. Unexpected end of file.
#document
| <html>
|   <head>
|   <body>
|     <cite>
|       <b>
|         <cite>
|           <i>
|             <cite>
|               <i>
|                 <cite>
|                   <i>
|       <i>
|         <i>
|           <div>
|             <b>
|               "X"
|             "TEST"
